inference_requirements.txt
fastapi==0.115.12
faiss-cpu==1.10.0
numpy==2.2.4
transformers==4.51.3
uvicorn==0.34.2